% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_implementation.R
\name{apply_dct}
\alias{apply_dct}
\title{apply_dct}
\usage{
apply_dct(x, type = "II")
}
\arguments{
\item{x}{representa o sinal}

\item{type}{representa o tipo de transformação DCT utilizada (type = 'I',
type = 'II', type = 'III' ou type = 'IV').}
}
\value{
retorna o sinal transformado pela DCT
}
\description{
Dado um vetor, realiza o cálculo da DCT, a implementação e baseada
nessa documentação do Matlab. No qual são utilizadas as DCT-I, DCT-II,
DCT-III e DCT-IV.
}
\details{
https://www.mathworks.com/help/signal/ref/dct.html
}
