% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mfcc.R
\name{apply_lifter}
\alias{apply_lifter}
\title{apply_lifter}
\usage{
apply_lifter(mfcc, cepstral.lifter = 22)
}
\arguments{
\item{mfcc}{uma matriz representando a MFCC, onde cada coluna representa os coeficientes da DCT.}

\item{cepstral.lifter}{consiste na quantidade de pontos calculados na FFT.}
}
\value{
Retorna a mfcc com a aplicação do sinusoidal liftering.
}
\description{
Esta função aplica o sinusoidal liftering aos MFCCs para realçar os MFCCs mais elevados.
E é definida como a seguinte equação:
\deqn{\hat{MFCC_i} = 1 + ((w_i * D)/2)sin((\pi * n)/D)}
}
