% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mfcc.R
\name{apply_preemphasis}
\alias{apply_preemphasis}
\title{apply_preemphasis}
\usage{
apply_preemphasis(x, a = 0.95)
}
\arguments{
\item{x}{refere-se a um vetor numérico representando o sinal de som.}

\item{a}{refere-se ao coeficiente do filtro \eqn{\alpha} , default, \eqn{\alpha}  = 0,95}
}
\value{
Retorna um vetor numérico representando o sinal após a aplicação do filtro
de pre-ênfase.
}
\description{
O filtro de pré-ênfase é utilizado com o objetivo de amplificar o sinal de
altas frequências e é util de várias maneiras como:
\enumerate{
\item equilibrar o espectro de frequências, uma vez que as frequências altas costumam ter magnitudes
menores em comparação com frequências mais baixas,
\item evitar problemas numéricos durante a operação de transformada de Fourier
\item também pode melhorar a relação sinal-ruido (SNR - Signal-to-Noise Ratio ).
}
O filtro pode ser aplicado usando a seguinte equação:
}
\details{
\deqn{y(t) = x(t) - \alpha x(t-1)}

os valores padrões do coeficiente \eqn{\alpha} são 0,95 ou 0,97
}
\examples{
x <- sample(50) #signal
print(x)
apply_preemphasis(x)
}
