% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfcc_implementation.R
\name{apply_window_hamming}
\alias{apply_window_hamming}
\title{apply_window_hamming}
\usage{
apply_window_hamming(frames)
}
\arguments{
\item{frames}{estrutura contendo os frames}
}
\value{
retorna a lista de frames com a aplicação do window hamming
}
\description{
Realiza a aplicação da função Window Hamming que é dada pela seguinte equação:
}
\details{
\deqn{w[n] = 0.56 - 0.46 * cos ((2\pi n)/N-1)}, onde o N representa o comprimento da janela.
}
