% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mfcc.R
\name{calcule_filter_bank}
\alias{calcule_filter_bank}
\title{calcule_filter_bank}
\usage{
calcule_filter_bank(m, f, npoints)
}
\arguments{
\item{m}{refere-se a um número real representando o filtro, este valor varia de 1 ate a quantidade
de filtros desejadas.}

\item{f}{um vetor de números reais representando as frequências convertidas de números de
bin fft.}

\item{npoints}{consiste na quantidade de pontos calculados na FFT.}
}
\value{
Retorna um vetor numérico representando o resultado do sinal sobre aquele filtro.
}
\description{
Esta função calcula o filter bank para um sinal aplicando um respectivo filtro \code{m}. Onde \code{m}
varia de 1 ate a quantidade filtros desejadas.
}
