% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfcc_implementation.R
\name{compute_mel_filterbanks}
\alias{compute_mel_filterbanks}
\title{compute_mel_filterbanks}
\usage{
compute_mel_filterbanks(freq.lower, freq.upper, num.filters, npoints,
  sample.rate)
}
\arguments{
\item{freq.lower}{valor mínimo da frequência dada em hertz}

\item{freq.upper}{valor maximo da frequência dada em hertz}

\item{num.filters}{número de filtros considerados para o cálculo}

\item{npoints}{número de pontos considerado para o cálculo dos N-Points FFT}

\item{sample.rate}{sample rate do sinal}
}
\value{
retorna a lista de frames com a aplicação do Power Spectrum
}
\description{
Realiza o cálculo dos mel-filterbank.
}
