% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfcc_implementation.R
\name{compute_power_spectrum}
\alias{compute_power_spectrum}
\title{compute_power_spectrum}
\usage{
compute_power_spectrum(frames, n = 512)
}
\arguments{
\item{frames}{estrutura contendo os frames}

\item{n}{valor representando o número de pontos para o cálculo da FFT}
}
\value{
retorna a lista de frames com a aplicação do Power Spectrum
}
\description{
Realiza a aplicação da função de Power Spectrum que é dada pela seguinte fórmula:
}
\details{
\deqn{(|FFT(X_i)|^2)/N}, onde o N representa o comprimento da janela.
}
