% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mfcc.R
\name{convert_hertz_to_mel}
\alias{convert_hertz_to_mel}
\title{convert_hertz_to_mel}
\usage{
convert_hertz_to_mel(hz)
}
\arguments{
\item{hz}{É um valor numérico que representa o valor da frequência em hertz}
}
\value{
Tem como retorno o valor em escala mel.
}
\description{
A escala Mel tem como objetivo imitar a percepção não-linear do ouvido humano, sendo
mais discriminativa nas freqüências mais baixas e menos discriminativa nas freqüências
mais altas. A conversão da frequência hertz para a escala mel é dada pela seguinte fórmula:
\deqn{Mel(hz) = 2595 * log10(1 + hz / 700)}
}
\examples{

mel <- convert_hertz_to_mel(1600)
print(mel)

}
