% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mfcc.R
\name{convert_mel_to_hertz}
\alias{convert_mel_to_hertz}
\title{convert_mel_to_hertz}
\usage{
convert_mel_to_hertz(mel)
}
\arguments{
\item{mel}{É um valor numérico que representa o valor da escala Mel}
}
\value{
A função retorna o valor numérico da frequência em hertz
}
\description{
Realiza a conversão da escala mel para a frequência em hertz. A fórmula é dada por:
\deqn{Hz(mel) = 700 * (10 ^ (mel / 2595) - 1)}
}
\examples{

hz <- convert_hertz_to_mel(1600)
print(hz)
}
