% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_implementation.R
\name{dct4}
\alias{dct4}
\title{dct4}
\usage{
dct4(k, x, N)
}
\arguments{
\item{k}{representa o valor usado para o kronecker delta}

\item{x}{representa o vetor numérico}

\item{N}{representa o valor que corresponde ao comprimento do vetor x.}
}
\value{
retorna o cálculo da DCT-4 para um vetor numérico.
}
\description{
O cálculo da DCT-4, é implementado conforme consta na documentação do matlab.
}
\details{
https://www.mathworks.com/help/signal/ref/dct.html
}
