% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mfcc.R
\name{filter_bank_function}
\alias{filter_bank_function}
\title{filter_bank_function}
\usage{
filter_bank_function(k, f, m)
}
\arguments{
\item{k}{refere-se a um valor real.}

\item{f}{um vetor de números reais representando as frequências convertidas de números de
bin fft.}

\item{m}{consiste em um número inteiro que varia no intervalo de 1 até os números de pontos
calculados na FFT .}
}
\value{
Retorna um valor numérico de acordo com o intervalo que se encontra o valor k.
}
\description{
Esta função realiza o cálculo sobre um valor numérico k de acordo com a fórmula definida na
seção \strong{Computing the Mel filterbank} do
seguinte \href{http://http://www.practicalcryptography.com/miscellaneous/machine-learning/guide-mel-frequency-cepstral-coefficients-mfccs//}{tutorial}.
}
