% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfcc_implementation.R
\name{frame_the_signal}
\alias{frame_the_signal}
\title{frame_the_signal}
\usage{
frame_the_signal(signal, sample.rate, frame.size = 0.025,
  frame.stride = 0.01)
}
\arguments{
\item{signal}{um vetor numérico, representando o sinal a ser dividido}

\item{sample.rate}{representa o sample rate do sinal}

\item{frame.size}{representa o comprimento da janela em segundos.}

\item{frame.stride}{representa o passo de cada janela}
}
\value{
retorna um dataframe, onde cada linha representa o trecho do sinal dividido
}
\description{
Realiza a divisão do sinal em short-frames
}
\examples{
x <- sample(1000)
frames <- frame_the_signal(x, 1600)
head(frames)

}
