% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_implementation.R
\name{kronecker_delta}
\alias{kronecker_delta}
\title{kronecker_delta}
\usage{
kronecker_delta(i, j = 0)
}
\arguments{
\item{i}{corresponde um valor inteiro.}

\item{j}{corresponde um valor inteiro.}
}
\value{
retorna um número de acordo com o valor i
}
\description{
A função kronecker delta e definida como:
}
\examples{
print(kronecker_delta(i = 1))

}
