% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mfcc.R
\docType{package}
\name{MFCC}
\alias{MFCC}
\alias{MFCC-package}
\title{MFCC: Mel-Frequency Cepstral Coefficients.}
\description{
Trabalho da disciplina Fundamentos de Computação Gráfica, implementação de
Mel-Frequency Cepstral Coefficients (MFCCs).
}
\references{
\itemize{
\item https://www.mathworks.com/help/signal/ref/dct.html
\item http://practicalcryptography.com/miscellaneous/machine-learning/guide-mel-frequency-cepstral-coefficients-mfccs/
\item https://docs.scipy.org/doc/numpy-1.14.0/reference/routines.fft.html
}
}
\author{
Jessica Cardoso
}
