% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfcc_implementation.R
\name{mfcc_function}
\alias{mfcc_function}
\title{mfcc_function}
\usage{
mfcc_function(signal, sample.rate, freq.lower, freq.upper, frame.size = 0.025,
  frame.stride = 0.01, nfft = 512, num.filters = 40, num.ceps = 13)
}
\arguments{
\item{signal}{vetor numérico representando o sinal}

\item{sample.rate}{sample rate do sinal}

\item{freq.lower}{valor mínimo da frequência dada em hertz}

\item{freq.upper}{valor maximo da frequência dada em hertz}

\item{frame.size}{representa o comprimento da janela em segundos.}

\item{frame.stride}{representa o passo de cada janela}

\item{nfft}{número de pontos considerado para o cálculo dos N-Points FFT}

\item{num.filters}{número de filtros considerados para o cálculo}

\item{num.ceps}{representa o número de coeficientes a serem mantidos}
}
\value{
retorna a lista de frames com a aplicação do Power Spectrum
}
\description{
Realiza o cálculo das MFFCs para um sinal.
}
