% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mfcc.R
\name{nfft}
\alias{nfft}
\title{nfft}
\usage{
nfft(signal, n = Inf)
}
\arguments{
\item{signal}{refere-se a um vetor numérico representando o sinal o qual se deseja
aplicar a transformada.}

\item{n}{parâmetro opcional que representa a quantidade de pontos a ser considerado para o
cálculo da DFT. Se nenhum valor for repassado, o cálculo será realizado sobre todos os elementos
do sinal. Se \code{n} for maior que o comprimento, o sinal e preenchido com zeros, se for menor,
o sinal e cortado.}
}
\value{
Retorna um vetor de comprimento teto de n/2, representando a aplicação da
transformada de Fourier sobre o sinal de entrada.
}
\description{
Esta função realiza o cálculo da Transformada Discreta de Fourier (DFT) sobre
n-pontos de um vetor unidimensional utilizando o algoritmo da Transformada
Rapida de Fourier (FFT). A DFT sobre valores reais tem como resultado um sinal
Hermitian symmetry, ou seja, os termos da frequência negativa sao apenas conjugados
complexos dos termos de frequência positiva. Os termos de valores negativos sao
descartados do resultado do cálculo para essa função.
}
\details{
Sobre DFT ser Hermitian symmetry:
https://docs.scipy.org/doc/numpy/reference/routines.fft.html
}
\examples{
x <- sample(50) #signal
print(x)
nfft(x)
}
