% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sound.data}
\alias{sound.data}
\title{Amostra de áudio para teste}
\format{An object of class \code{Wave} of length 28000.}
\source{
\url{http://www.diamondse.info/}
}
\usage{
sound.data
}
\description{
Amostra de áudio para teste
}
\keyword{datasets}
